fix(core): satisfy R2 complexity budget
Fold the private stdio host into the existing Assignment Runtime package, keep coordinator readiness local to its runtime service owner, and refresh the portable source projection.
